+++ src/main.tsx
@@ -25,6 +25,7 @@ import { TagService } from '@/services/tag_service';
 import { AutoCompleteSuggest } from '@/views/auto_completete_suggest';
 import { registerRightClickMenu } from '@/ui/rightButtonMenu';
 import { LineHashCache } from '@/cache/lineHashCache';
+import { WidgetSizeManager } from './settings/widgetSizeManager';
 
 
 export default class EquationCitator extends Plugin {
@@ -45,6 +46,7 @@ export default class EquationCitator extends Plugin {
     async onload() {
         await this.loadSettings();
         ColorManager.updateAllColors(this.settings);
+        WidgetSizeManager.updateAllSizes(this.settings);
         this.addSettingTab(new SettingsTabView(this.app, this));
         // initialize caches
         this.loadCaches();
@@ -59,7 +61,7 @@ export default class EquationCitator extends Plugin {
         this.loadEditorExtensions();
         this.loadReadingModeExtensions();
         this.registerEditorSuggest(this.autoCompleteSuggest);
-        
+
         // register ribbon button and commands 
         registerRibbonButton(this);
         registerRightClickMenu(this);